separate sockets from main server file
