+++ src/server.js
@@ -8,6 +8,7 @@ const cors = require('cors');
 const multer = require('multer');
 const uploadConfig = require('./config/upload.js');
 const upload = multer(uploadConfig);
+const websockets = require('./sockets.js');
 
 const app = express();
 app.use(cors());
@@ -15,71 +16,9 @@ app.use(express.json())
 
 const server = http.createServer(app);
 const sockets = new Server(server);
-
-const rooms = {};
-const usersSpot = {};
-
-function isUserIn(userId, room){
-    const user = rooms[room].users.find(item=>item === userId);
-
-    return !!user;
-}
-
-function joinUser(userId, room){
-    if(isUserIn(userId,room)) return;
-    removeUserFromTheLastRoom(userId);
-
-    usersSpot[userId] = room;
-    rooms[room].users.push(userId);
-    getMessage(room);
-}
-
-function createRoom(userId, room){
-    removeUserFromTheLastRoom(userId);
-     rooms[room] = [userId];
-     rooms[room] = {
-        users:[userId],
-        text:''
-     }
-     usersSpot[userId] = room;
-}
-
-function getMessage(room){
-    rooms[room].users.map(id=>{
-        sockets.to(id).emit('message', rooms[room].text);
-    })
-}
-
-function removeUserFromTheLastRoom(userId){
-    if(!usersSpot[userId]) return;
-
-    const room = usersSpot[userId]; 
-
-    const index = rooms[room].users.indexOf(userId);
-    rooms[room].users.splice(index,1);
-}
-
+websockets(sockets);
 
 app.use( express.static(path.join(__dirname,'public')) );
-
-sockets.on('connection', (socket)=>{
-    const userId = socket.id; 
-
-    socket.on('transferData', (roomData)=>{
-        const { room, text } = roomData;
-        rooms[room].text = text;
-
-        getMessage(room);
-    })
-
-    socket.on('join', (room)=>{
-        rooms[room] ? joinUser(userId, room) : createRoom(userId, room);
-    })
-
-    socket.on('disconnect', ()=>{
-        removeUserFromTheLastRoom(userId);
-    })
-})
 
 app.post("/upload", upload.single("image"),(req,res)=>{
     return res.status(200).json({ message: "Your image has been uploaded!"});
